requirements.txt
Django==5.1.2
django-bootstrap5==24.3
djangorestframework==3.15.2
drf-nested-routers==0.94.1